Move test resources under parsing module
