+++ service-capabilities/pom.xml
@@ -17,6 +17,11 @@
             <groupId>org.oskari</groupId>
             <artifactId>service-base</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.oskari</groupId>
+            <artifactId>shared-test-resources</artifactId>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 </project>
