+++ service-capabilities/src/test/resources/org/oskari/capabilities/ogc/WMTSCapabilitiesParserTest-LMI-input.xml
