+++ service-capabilities/src/test/resources/org/oskari/capabilities/ogc/WMTSCapabilitiesParserTest-NLS-input.xml
