+++ service-capabilities/src/test/resources/org/oskari/capabilities/ogc/WMTSCapabilitiesParserTest-asdi-input.xml
