+++ service-capabilities/src/test/resources/org/oskari/capabilities/ogc/WMTSCapabilitiesParserTest-dummy-expected.json
