+++ service-capabilities/src/test/resources/org/oskari/capabilities/ogc/WMTSCapabilitiesParserTest-spain-input.xml
