--- service-map/src/test/resources/fi/nls/oskari/wmts/WMTSCapabilitiesParserTest-expected-single-layer-NLS-taustakartta.json
@@ -1,215 +1,0 @@
-{
-  "id": "taustakartta",
-  "title": "Taustakartta",
-  "styles": [
-    {
-      "title": "default",
-      "legend": null,
-      "name": "default"
-    }
-  ],
-  "defaultStyle": "default",
-  "formats": [
-    "image/png"
-  ],
-  "infoFormats": [],
-  "resourceUrls": [
-    {
-      "format": "image/png",
-      "type": "tile",
-      "template": "http://karttamoottori.maanmittauslaitos.fi/maasto/wmts/1.0.0/taustakartta/default/{TileMatrixSet}/{TileMatrix}/{TileRow}/{TileCol}.png"
-    }
-  ],
-  "links": [
-    {
-      "limits": null,
-      "tileMatrixSet": {
-        "identifier": "ETRS-TM35FIN",
-        "projection": "urn:ogc:def:crs:EPSG:6.3:3067",
-        "matrixIds": {
-          "11": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 2048,
-            "matrixHeight": 2048,
-            "identifier": "11",
-            "scaleDenominator": 14285.714285714286,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          },
-          "12": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 4096,
-            "matrixHeight": 4096,
-            "identifier": "12",
-            "scaleDenominator": 7142.857142857143,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          },
-          "13": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 8192,
-            "matrixHeight": 8192,
-            "identifier": "13",
-            "scaleDenominator": 3571.4285714285716,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          },
-          "14": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 16384,
-            "matrixHeight": 16384,
-            "identifier": "14",
-            "scaleDenominator": 1785.7142857142858,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          },
-          "0": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 1,
-            "matrixHeight": 1,
-            "identifier": "0",
-            "scaleDenominator": 2.925714285714286E7,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          },
-          "1": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 2,
-            "matrixHeight": 2,
-            "identifier": "1",
-            "scaleDenominator": 1.462857142857143E7,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          },
-          "2": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 4,
-            "matrixHeight": 4,
-            "identifier": "2",
-            "scaleDenominator": 7314285.714285715,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          },
-          "3": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 8,
-            "matrixHeight": 8,
-            "identifier": "3",
-            "scaleDenominator": 3657142.8571428573,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          },
-          "4": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 16,
-            "matrixHeight": 16,
-            "identifier": "4",
-            "scaleDenominator": 1828571.4285714286,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          },
-          "5": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 32,
-            "matrixHeight": 32,
-            "identifier": "5",
-            "scaleDenominator": 914285.7142857143,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          },
-          "6": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 64,
-            "matrixHeight": 64,
-            "identifier": "6",
-            "scaleDenominator": 457142.85714285716,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          },
-          "7": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 128,
-            "matrixHeight": 128,
-            "identifier": "7",
-            "scaleDenominator": 228571.42857142858,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          },
-          "8": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 256,
-            "matrixHeight": 256,
-            "identifier": "8",
-            "scaleDenominator": 114285.71428571429,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          },
-          "9": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 512,
-            "matrixHeight": 512,
-            "identifier": "9",
-            "scaleDenominator": 57142.857142857145,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          },
-          "10": {
-            "tileWidth": 256,
-            "tileHeight": 256,
-            "matrixWidth": 1024,
-            "matrixHeight": 1024,
-            "identifier": "10",
-            "scaleDenominator": 28571.428571428572,
-            "topLeftCorner": {
-              "lon": -548576.0,
-              "lat": 8388608.0
-            }
-          }
-        }
-      }
-    }
-  ],
-  "queryable": false
-}
